# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "charpente API",
    "description": "Backend API. Errors are RFC 9457 problem+json with a stable `code`.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/uploads": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "The caller's uploads, newest first, cursor-paginated.",
        "operationId": "list_uploads",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1..=100 (default 25).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UploadDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      },
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Declare an upload: validates type + size, records a pending row, returns a\npresigned PUT URL.",
        "operationId": "create_upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/api/v1/uploads/{id}/complete": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Confirm the browser finished its PUT; verifies the object actually exists.",
        "operationId": "complete_upload",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Upload id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDto"
                }
              }
            }
          },
          "404": {
            "description": "No such pending upload"
          },
          "409": {
            "description": "Object not found in storage"
          }
        }
      }
    },
    "/api/v1/uploads/{id}/download": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "A short-lived download URL for a completed upload.",
        "operationId": "download_upload",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Upload id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadUploadResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such completed upload"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateUploadRequest": {
        "type": "object",
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CreateUploadResponse": {
        "type": "object",
        "required": [
          "upload",
          "put_url"
        ],
        "properties": {
          "put_url": {
            "type": "string",
            "description": "Presigned URL: PUT the file bytes here with the declared content type."
          },
          "upload": {
            "$ref": "#/components/schemas/UploadDto"
          }
        }
      },
      "DownloadUploadResponse": {
        "type": "object",
        "required": [
          "get_url"
        ],
        "properties": {
          "get_url": {
            "type": "string",
            "description": "Presigned URL, valid for a few minutes."
          }
        }
      },
...
      "Page_UploadDto": {
        "type": "object",
        "required": [
          "items",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "filename",
                "content_type",
                "size_bytes",
                "state",
                "created_at"
              ],
              "properties": {
                "content_type": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "filename": {
                  "type": "string"
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "size_bytes": {
                  "type": "integer",
                  "format": "int64"
                },
                "state": {
                  "$ref": "#/components/schemas/UploadState"
                }
              }
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Pass as `cursor` to fetch the next page; absent on the last page."
          }
        }
      },
...
      "UploadDto": {
        "type": "object",
        "required": [
          "id",
          "filename",
          "content_type",
          "size_bytes",
          "state",
          "created_at"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "filename": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "state": {
            "$ref": "#/components/schemas/UploadState"
          }
        }
      },
      "UploadState": {
        "type": "string",
        "enum": [
          "pending",
          "complete"
        ]
      },
...
    }
  },
...
}